+++ moe/bandit/epsilon_first.py
@@ -1,15 +1,16 @@
 # -*- coding: utf-8 -*-
 """Classes (Python) to compute the Bandit Epsilon-First arm allocation and choosing the arm to pull next.
 
-See :class:`moe.bandit.epsilon.Epsilon` for further details on bandit.
+See :class:`moe.bandit.epsilon_interface.EpsilonInterface` for further details on bandit.
 
 """
 from moe.bandit.constant import DEFAULT_EPSILON, DEFAULT_TOTAL_SAMPLES, EPSILON_SUBTYPE_FIRST
-from moe.bandit.epsilon import Epsilon
+from moe.bandit.epsilon_interface import EpsilonInterface
+
 from moe.bandit.utils import get_equal_arm_allocations
 
 
-class EpsilonFirst(Epsilon):
+class EpsilonFirst(EpsilonInterface):
 
     r"""Implementation of EpsilonFirst.
 
@@ -19,7 +20,7 @@ class EpsilonFirst(Epsilon):
     number sampled is calculated by summing up total from each arm sampled.
     total_samples is T from :doc:`bandit`.
 
-    See superclass :class:`moe.bandit.epsilon.Epsilon` for further details.
+    See superclass :class:`moe.bandit.epsilon_interface.EpsilonInterface` for further details.
 
     """
 
@@ -29,7 +30,7 @@ class EpsilonFirst(Epsilon):
             epsilon=DEFAULT_EPSILON,
             total_samples=DEFAULT_TOTAL_SAMPLES,
     ):
-        """Construct an EpsilonFirst object. See superclass :class:`moe.bandit.epsilon.Epsilon` for details.
+        """Construct an EpsilonFirst object. See superclass :class:`moe.bandit.epsilon_interface.EpsilonInterface` for details.
 
         total_samples is the total number of samples (number to sample + number sampled)
         number sampled is calculated by summing up total from each arm sampled.