+++ moe/tests/bandit/ucb1_tuned_test.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""Test UCB1 bandit implementation.
+"""Test UCB1-tuned bandit implementation.
 
 Test default values with one, two, and three arms.
 Test different cases including unsampled arms and multiple winners.
@@ -7,15 +7,15 @@ Test different cases including unsampled arms and multiple winners.
 """
 import testify as T
 
-from moe.bandit.ucb1 import UCB1
+from moe.bandit.ucb1_tuned import UCB1Tuned
 from moe.tests.bandit.ucb_test_case import UCBTestCase
 
 
-class UCB1Test(UCBTestCase):
+class UCB1TunedTest(UCBTestCase):
 
     """Verify that different historical infos return correct results."""
 
-    bandit_class = UCB1
+    bandit_class = UCB1Tuned
 
     def test_init_default(self):
         """Verify that default values do not throw and error. This is purely an integration test."""
@@ -37,6 +37,11 @@ class UCB1Test(UCBTestCase):
     def test_three_arms_two_winners(self):
         """Check that the three-arms cases with two winners return the expected arm allocations. This tests num_arms > num_winning_arms > 1."""
         self._test_three_arms_two_winners()
+
+    def test_three_arms_diferent_variance(self):
+        """Check that the three-arms cases with different variance (same average payoff) return the expected arm allocations. The highest variance wins."""
+        bandit = self.bandit_class(self.three_arms_with_variance_no_unsampled_arm_test_case)
+        T.assert_dicts_equal(bandit.allocate_arms(), {"arm1": 1.0, "arm2": 0.0, "arm3": 0.0})
 
 
 if __name__ == "__main__":